requirements.txt
typarse==3.4.0
numpy>=1.23.5
gymnasium~=0.27.1
jupyter>=1.0.0
jupyterlab>=3.5.3
mlagents-envs~=0.28.0
PyYAML>=6.0
scipy>=1.10.0
matplotlib==3.6.3
seaborn>=0.12.2
tensorboard>=2.11.2
torch>=1.13.1
tqdm>=4.64.1
ipykernel>=6.20.2
numba>=0.56.4
pytest>=7.2.1
coverage>=7.1.0
wandb>=0.13.9
pybullet==3.2.5
opencv-python~=4.7.0.68
PettingZoo[sisl]>=1.22.3
supersuit>=3.7.1
cloudpickle~=2.2.1
pillow~=9.4.0
setuptools~=65.6.3
pyvirtualdisplay~=3.0
optuna~=3.1.0
pytype==2023.1.17
jax==0.4.1
jaxlib==0.4.1
shortuuid==1.0.11
black==22.3.0